Split files and add break
