+++ shell-language/examples/break.sh
@@ -1,0 +1,6 @@
+let s = literal "abcde"
+for split $s "" each
+	echo "Char: $iter"
+	set break = if_equal $iter "c" break ""
+
+echo $iter
