+++ shell-language/src/main.rs
@@ -1,12 +1,16 @@
+pub mod evaluate;
+pub mod parsing;
+pub mod utilities;
+
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     use std::{env, fs};
 
     let mut args = env::args().skip(1);
     let Some(first) = args.next() else {
-        let run_id = option_env!("GITHUB_RUN_ID");
-        let date = option_env!("GIT_LAST_COMMIT");
-        let after = run_id
-            .map(|commit| format!(" (commit {commit} {date:?})"))
+        let commit = option_env!("GIT_LAST_COMMIT");
+        let date = option_env!("GITHUB_RUN_DATE").unwrap_or_default();
+        let after = commit
+            .map(|commit| format!(" (commit {commit} {date})"))
             .unwrap_or_default();
         println!("the Ben shell (WIP){after}");
         return Ok(());
@@ -54,7 +58,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         let rest: Vec<_> = args.collect();
 
         let debug_program: bool = rest.iter().any(|flag| flag == "--debug-program");
-        let program = parsing::parse_program(&source);
+        let program = parsing::parsing::parse_program(&source);
 
         if debug_program {
             eprintln!("{program:#?}");
@@ -66,903 +70,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
-mod ast {
-    #[derive(Debug)]
-    pub struct Program<'a>(pub Vec<Statement<'a>>);
-
-    #[derive(Debug)]
-    pub enum Statement<'a> {
-        Declaration {
-            name: &'a str,
-            value: Command<'a>,
-        },
-        For {
-            iterator: Command<'a>,
-            statements: Vec<Statement<'a>>,
-        },
-        Command(Command<'a>),
-    }
-
-    #[derive(Debug)]
-    pub struct Command<'a> {
-        pub name: &'a str,
-        pub arguments: Vec<Argument<'a>>,
-    }
-
-    /// Holds strings and literals
-    #[derive(Debug)]
-    pub struct Argument<'a>(pub &'a str);
-}
-
-mod parsing {
-    use super::ast::{Argument, Command, Program, Statement};
-
-    pub type Lines<'a> = std::iter::Peekable<std::str::Lines<'a>>;
-
-    pub fn parse_program(on: &str) -> Program<'_> {
-        let mut stmts: Vec<Statement> = Vec::new();
-
-        let mut lines = on.lines().peekable();
-        while let Some(line) = lines.next() {
-            if let Some(stmt) = parse_statement(line, &mut lines, 0) {
-                stmts.push(stmt);
-            }
-        }
-        Program(stmts)
-    }
-
-    pub fn strip_ident(mut line: &str, upto: usize) -> &str {
-        for _ in 0..upto {
-            line = line
-                .strip_prefix("\t")
-                .or_else(|| line.strip_prefix("  "))
-                .unwrap_or(line);
-        }
-        line
-    }
-
-    pub fn parse_statement<'a>(
-        line: &'a str,
-        lines: &mut Lines<'a>,
-        depth: usize,
-    ) -> Option<Statement<'a>> {
-        let line = strip_ident(line, depth);
-        // comment or empty
-        if line.starts_with('#') || line.trim().is_empty() {
-            None
-        } else if let Some(rest) = line.trim_start().strip_prefix("let ") {
-            let (name, rest) = rest.split_once(" = ").expect("let declaration needs ' = '");
-            Some(Statement::Declaration {
-                name,
-                value: parse_command(rest),
-            })
-        } else if let Some(inner) = line
-            .trim_start()
-            .strip_prefix("for ")
-            .and_then(|line| line.strip_suffix(" each"))
-        {
-            let iterator = parse_command(inner);
-            let mut statements = Vec::new();
-            loop {
-                let line = lines.next().expect("expected for loop body");
-                if let Some(stmt) = parse_statement(line, lines, depth + 1) {
-                    statements.push(stmt);
-                }
-                let r#continue = lines
-                    .peek()
-                    .map(|line| strip_ident(line, depth))
-                    .is_some_and(|line: &str| {
-                        line.is_empty() || line.starts_with("\t") || line.starts_with("  ")
-                    });
-                if !r#continue {
-                    break;
-                }
-            }
-            Some(Statement::For {
-                iterator,
-                statements,
-            })
-        } else {
-            Some(Statement::Command(parse_command(line)))
-        }
-    }
-
-    fn parse_command(on: &str) -> Command<'_> {
-        let mut name = "";
-        let mut arguments = Vec::new();
-        let mut in_string: Option<char> = None;
-        let mut last = 0;
-        let mut escaped = false;
-        for (idx, chr) in on.char_indices() {
-            if let Some(matcher) = in_string {
-                if escaped {
-                    escaped = false;
-                    continue;
-                }
-                if chr == matcher {
-                    // arguments.push(Argument(&on[last..=idx]));
-                    // last = idx + 1;
-                    in_string = None;
-                }
-                escaped = chr == '\\';
-            } else if let ('"' | '\'' | '`', "") = (chr, on[last..idx].trim()) {
-                in_string = Some(chr);
-                // last = idx;
-            } else if let ' ' = chr {
-                let part = on[last..idx].trim();
-                if !part.is_empty() {
-                    if name.is_empty() {
-                        name = part;
-                    } else {
-                        arguments.push(Argument(part));
-                    }
-                    last = idx + chr.len_utf8();
-                }
-            }
-        }
-        let rest = &on[last..].trim();
-        if !rest.is_empty() {
-            if name.is_empty() {
-                name = rest;
-            } else {
-                arguments.push(Argument(rest));
-            }
-        }
-        Command { name, arguments }
-    }
-}
-
-mod evaluate {
-    use super::ast::{Argument, Command, Program, Statement};
-
-    use std::borrow::Cow;
-    use std::collections::HashMap;
-    use std::fs;
-
-    pub type Context<'a> = HashMap<&'a str, String>;
-
-    pub fn evaluate_program(program: &Program<'_>) {
-        let mut ctx: Context<'_> = HashMap::new();
-        for statement in &program.0 {
-            evaluate_statement(statement, &mut ctx);
-        }
-    }
-
-    pub fn evaluate_statement<'a>(statement: &Statement<'a>, ctx: &mut Context<'a>) {
-        match statement {
-            Statement::Declaration { name, value } => {
-                let (value, exit_code) = evaluate_command(value, ctx);
-                ctx.insert(name, value);
-                if let Some(exit_code) = exit_code {
-                    ctx.insert("exit_code", exit_code.to_string());
-                }
-            }
-            Statement::Command(command) => {
-                let (_, exit_code) = evaluate_command(command, ctx);
-                if let Some(exit_code) = exit_code {
-                    ctx.insert("exit_code", exit_code.to_string());
-                }
-            }
-            Statement::For {
-                iterator,
-                statements,
-            } => {
-                let (result, exit_code) = evaluate_command(iterator, ctx);
-                if let Some(exit_code) = exit_code {
-                    ctx.insert("exit_code", exit_code.to_string());
-                }
-
-                for part in result.trim_end().split('\n') {
-                    let part = part.strip_suffix('\r').unwrap_or(part);
-                    match iterator.name {
-                        "git" => match iterator.arguments.first().map(|arg| arg.0) {
-                            Some("tag") => {
-                                ctx.insert("tag", part.to_owned());
-                            }
-                            Some("log") => {
-                                ctx.insert("ref", part.to_owned());
-                            }
-                            _ => {}
-                        },
-                        "files" => {
-                            ctx.insert("file", part.to_owned());
-                        }
-                        "constant" => {
-                            let first_argument = iterator.arguments[0].0;
-                            if let Some(name) = first_argument
-                                .strip_prefix('$')
-                                .and_then(|rest| crate::utilities::depluralise(rest))
-                            {
-                                ctx.insert(name, part.to_owned());
-                            }
-                        }
-                        _ => {}
-                    }
-                    ctx.insert("iter", part.to_owned());
-                    for statement in statements {
-                        evaluate_statement(statement, ctx);
-                    }
-                }
-            }
-        }
-    }
-
-    /// interpolate variables
-    fn evaluate_argument<'a>(argument: &Argument<'a>, ctx: &'a Context<'a>) -> Cow<'a, str> {
-        let mut result = Cow::Borrowed("");
-        let mut start = 0;
-        let on = &argument.0;
-        let mut last_was_escape_backslash = false;
-        for (idx, matched) in on.match_indices(['$', '\\', '\'', '"']) {
-            let skip = last_was_escape_backslash && matched == "\\";
-            last_was_escape_backslash = false;
-            if skip {
-                continue;
-            }
-
-            result += &on[start..idx];
-            if let "$" = matched {
-                let rest = &on[(idx + 1)..];
-                let reference = rest
-                    .split_once(|chr: char| !(chr.is_alphanumeric() || matches!(chr, '_')))
-                    .map_or(rest, |(rest, _)| rest);
-                if let "ctx" = reference {
-                    result += Cow::Owned(format!("{ctx:?}"));
-                } else if let Some(argument) = ctx.get(&reference) {
-                    result += Cow::Borrowed(argument.as_str());
-                } else if let Some(env) = crate::utilities::get_environment_variable(reference) {
-                    result += Cow::Owned(env);
-                } else {
-                    eprintln!("shell-language: Could not find reference {reference}");
-                }
-                start = idx + 1 + reference.len();
-            } else if let "\\" = matched {
-                match on[(idx + 1)..].chars().next() {
-                    Some('n') => {
-                        result += Cow::Borrowed("\n");
-                    }
-                    Some('t') => {
-                        result += Cow::Borrowed("\t");
-                    }
-                    Some('r') => {
-                        result += Cow::Borrowed("\r");
-                    }
-                    Some('\\') => {
-                        last_was_escape_backslash = true;
-                        result += Cow::Borrowed("\\");
-                    }
-                    Some('"') => {
-                        result += Cow::Borrowed("\"");
-                    }
-                    Some('\'') => {
-                        result += Cow::Borrowed("'");
-                    }
-                    character => {
-                        eprintln!("unknown escape {character:?}");
-                    }
-                }
-                start = idx + 2;
-            } else if let "\"" | "'" = matched {
-                let skip = on[..idx].is_empty()
-                    || on[..idx].ends_with(&['=', '\\'])
-                    || on[idx..][1..].is_empty();
-                start = if skip { idx + 1 } else { idx };
-            } else {
-                unreachable!("matched '{matched}'");
-            }
-        }
-        result += &on[start..];
-        result
-    }
-
-    #[allow(clippy::too_many_lines)]
-    pub fn evaluate_command(command: &Command<'_>, ctx: &Context) -> (String, Option<i32>) {
-        match command.name {
-            // Command line printing
-            "echo" | "echo_stdout" => {
-                let mut some = false;
-                if let Some("run") = command.arguments.first().map(|arg| arg.0) {
-                    let mut arguments = command.arguments[1..].iter();
-
-                    let first_argument = arguments.next().expect("command name");
-                    let command = evaluate_argument(first_argument, ctx);
-                    let args = arguments
-                        .map(|arg| evaluate_argument(arg, ctx).into_owned())
-                        .filter(|arg| !arg.is_empty())
-                        .collect::<Vec<String>>();
-
-                    let (_output, result) = crate::utilities::run_command(
-                        command.into_owned(),
-                        args,
-                        None,
-                        false,
-                        false,
-                    );
-
-                    (String::new(), result.code())
-                } else {
-                    for (idx, argument) in command.arguments.iter().enumerate() {
-                        let result = evaluate_argument(argument, ctx);
-                        if !result.is_empty() {
-                            some = true;
-                            if idx > 0 {
-                                print!(" ");
-                            }
-                            print!("{result}");
-                        }
-                    }
-                    if command.arguments.is_empty() || some {
-                        println!();
-                    }
-                    (String::new(), None)
-                }
-            }
-            "echo_stderr" => {
-                for (idx, argument) in command.arguments.iter().enumerate() {
-                    if idx > 0 {
-                        eprint!(" ");
-                    }
-                    eprint!("{}", evaluate_argument(argument, ctx));
-                }
-                eprintln!();
-                (String::new(), None)
-            }
-            // Run command
-            name @ ("run" | "with") => {
-                let mut arguments = command.arguments.iter();
-                let mut env: Vec<(String, String)> = Vec::new();
-                if let "with" = name {
-                    while let Some(key) = arguments.next() {
-                        if let "run" = key.0 {
-                            break;
-                        }
-                        let key = evaluate_argument(key, ctx);
-                        let value = arguments.next().expect("env value");
-                        let value = evaluate_argument(value, ctx);
-                        if !value.is_empty() {
-                            env.push((key.into_owned(), value.into_owned()));
-                        }
-                    }
-                }
-
-                let first_argument = arguments.next().expect("command name");
-                let command = evaluate_argument(first_argument, ctx);
-                let mut args: Vec<String> = arguments
-                    .map(|arg| evaluate_argument(arg, ctx).into_owned())
-                    .filter(|arg| !arg.is_empty())
-                    .collect();
-
-                let (capture_stdout, capture_stderr) = if args
-                    .pop_if(|top| top == "--merge-stdout-and-stderr")
-                    .is_some()
-                {
-                    (true, true)
-                } else if args.pop_if(|top| top == "--only-capture-stderr").is_some() {
-                    (false, true)
-                } else {
-                    (true, false)
-                };
-
-                let (output, result) = crate::utilities::run_command(
-                    command.into_owned(),
-                    args,
-                    Some(env),
-                    capture_stdout,
-                    capture_stderr,
-                );
-
-                (output, result.code())
-            }
-            // Environment variables
-            "env" => {
-                let mut arguments = command.arguments.iter();
-                let name: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                if let Some(value) = crate::utilities::get_environment_variable(name) {
-                    (value, Some(0))
-                } else {
-                    eprintln!("Could not find environment variable {name}");
-                    (String::default(), Some(1))
-                }
-            }
-            // Filesystem manipulation
-            "mv" | "move" => {
-                use std::path::Path;
-
-                let mut arguments = command.arguments.iter();
-                let from: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let to: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let response =
-                    crate::utilities::move_copy_file(Path::new(from), Path::new(to), true);
-                match response {
-                    Ok(()) => (String::default(), Some(0)),
-                    Err(err) => {
-                        eprintln!("error moving file: {err:?}");
-                        (String::default(), Some(0))
-                    }
-                }
-            }
-            "cp" | "copy" => {
-                use std::path::Path;
-
-                let mut arguments = command.arguments.iter();
-                let from: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let to: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-
-                let response =
-                    crate::utilities::move_copy_file(Path::new(from), Path::new(to), false);
-                match response {
-                    Ok(()) => (String::default(), Some(0)),
-                    Err(err) => {
-                        eprintln!("error copying file: {err:?}");
-                        (String::default(), Some(0))
-                    }
-                }
-            }
-            "rm" | "remove" => {
-                use std::path::Path;
-
-                let mut arguments = command.arguments.iter();
-                let path: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let path: &Path = Path::new(path);
-                if path.is_dir() {
-                    fs::remove_dir(path).unwrap();
-                    (String::default(), Some(0))
-                } else if path.is_file() {
-                    fs::remove_file(path).unwrap();
-                    (String::default(), Some(0))
-                } else {
-                    eprintln!("unknown path item to remove");
-                    (String::default(), Some(1))
-                }
-            }
-            // Scan files
-            "files" => {
-                let pattern = if let Some(arg) = command.arguments.first() {
-                    evaluate_argument(arg, ctx)
-                } else {
-                    Cow::Borrowed("")
-                };
-                match glob::glob(&pattern) {
-                    Ok(paths) => {
-                        let mut output = String::new();
-                        for path in paths {
-                            if !output.is_empty() {
-                                output.push('\n');
-                            }
-                            output
-                                .push_str(&path.unwrap().display().to_string().replace('\\', "/"));
-                        }
-                        (output, Some(0))
-                    }
-                    Err(err) => {
-                        eprintln!("Error reading files glob {err:?}");
-                        (String::default(), Some(1))
-                    }
-                }
-            }
-            // File reads and writes
-            "write" => {
-                let mut arguments = command.arguments.iter();
-                let path: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let output: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                if fs::write(path, output).is_ok() {
-                    (String::default(), Some(0))
-                } else {
-                    eprintln!("Could not write to {path}");
-                    (String::default(), Some(1))
-                }
-            }
-            "read" => {
-                let mut arguments = command.arguments.iter();
-                let path: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                if let Ok(content) = fs::read_to_string(path) {
-                    (content, Some(0))
-                } else {
-                    eprintln!("Could not read {path}");
-                    (String::default(), Some(1))
-                }
-            }
-            "append" => {
-                let mut arguments = command.arguments.iter();
-                let path: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let to_append: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                if let Ok(mut content) = fs::read_to_string(path) {
-                    // I think this is okay
-                    content.push('\n');
-                    content.push_str(to_append);
-                    if fs::write(path, content).is_ok() {
-                        (String::default(), Some(0))
-                    } else {
-                        eprintln!("Could not write to {path}");
-                        (String::default(), Some(1))
-                    }
-                } else {
-                    eprintln!("Could not read {path}");
-                    (String::default(), Some(1))
-                }
-            }
-            // String commands
-            "repeat" => {
-                let mut arguments = command.arguments.iter();
-                let item: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let repeat: usize = evaluate_argument(arguments.next().unwrap(), ctx)
-                    .parse()
-                    .expect("invalid repeater");
-                (item.repeat(repeat), None)
-            }
-            "replace" => {
-                let mut arguments = command.arguments.iter();
-                let item: &str = &evaluate_argument(arguments.next().expect("no item"), ctx);
-                let from: &str =
-                    &evaluate_argument(arguments.next().expect("no item to replace"), ctx);
-                let to: &str = &evaluate_argument(arguments.next().expect("no replacer"), ctx);
-                (item.replace(from, to), None)
-            }
-            "concatenate" => {
-                use std::fmt::Write;
-
-                let mut s = String::new();
-                for argument in &command.arguments {
-                    let argument = evaluate_argument(argument, ctx);
-                    if !argument.is_empty() {
-                        if !s.is_empty() {
-                            writeln!(&mut s).unwrap();
-                        }
-                        write!(&mut s, "{argument}").unwrap();
-                    }
-                }
-                (s, None)
-            }
-            "concatenate_separator" => {
-                // could concatenate + replace...
-                use std::fmt::Write;
-
-                let mut s = String::new();
-                let mut arguments = command.arguments.iter();
-                let separator: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                for argument in arguments {
-                    let argument = evaluate_argument(argument, ctx);
-                    if !argument.is_empty() {
-                        if !s.is_empty() {
-                            write!(&mut s, "{separator}").unwrap();
-                        }
-                        write!(&mut s, "{argument}").unwrap();
-                    }
-                }
-                (s, None)
-            }
-            str_slice_cmd @ ("before" | "after" | "rbefore" | "rafter") => {
-                let mut arguments = command.arguments.iter();
-                let item: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let splitter: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-
-                let item = if str_slice_cmd.starts_with('r') {
-                    item.rsplit_once(splitter)
-                } else {
-                    item.split_once(splitter)
-                };
-                let out = if let Some((before, after)) = item {
-                    if str_slice_cmd.ends_with("before") {
-                        before.to_owned()
-                    } else {
-                        after.to_owned()
-                    }
-                } else {
-                    arguments
-                        .next()
-                        .map(|default| evaluate_argument(default, ctx).into_owned())
-                        .unwrap_or_default()
-                };
-                (out, None)
-            }
-            line_cmd @ ("last_line" | "first_line") => {
-                let mut arguments = command.arguments.iter();
-                let first_argument = arguments.next().unwrap();
-                let item: &str = &evaluate_argument(first_argument, ctx);
-
-                let mut lines = item.lines();
-                let out = if line_cmd.starts_with("first") {
-                    lines.next()
-                } else {
-                    lines.next_back()
-                };
-                (out.unwrap_or_default().to_owned(), None)
-            }
-            "size" => {
-                let mut arguments = command.arguments.iter();
-                let item: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                (item.len().to_string(), None)
-            }
-            "lines" => {
-                let mut arguments = command.arguments.iter();
-                let item: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                (item.lines().count().to_string(), None)
-            }
-            "trim" => {
-                let mut arguments = command.arguments.iter();
-                let item: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                (item.trim().to_owned(), None)
-            }
-            "format_number" => {
-                let mut arguments = command.arguments.iter();
-                let item: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                // implementation is a little borked
-                let result = if item.contains('.') {
-                    let num: f64 = item.parse().expect("cannot format non float");
-                    let is_negative = num.is_sign_negative();
-                    let abs_num = num.abs();
-                    let (whole, fract) = (abs_num.trunc(), abs_num.fract());
-                    let sign = if is_negative { "-" } else { "" };
-                    let whole = crate::utilities::separate_numbers(&whole.to_string());
-                    let fract = if fract == 0.0 {
-                        String::new()
-                    } else {
-                        let fract =
-                            crate::utilities::separate_numbers_fract(&fract.to_string()[2..]);
-                        format!(" . {fract}")
-                    };
-                    format!("{sign}{whole}{fract}")
-                } else if item.contains('-') {
-                    let num: i64 = item.parse().expect("cannot format non integer");
-                    let is_negative = num.is_negative();
-                    let sign = if is_negative { "-" } else { "" };
-                    let whole = crate::utilities::separate_numbers(&num.abs().to_string());
-                    format!("{sign}{whole}")
-                } else {
-                    let num: u64 = item.parse().expect("cannot format non natural");
-                    crate::utilities::separate_numbers(&num.to_string())
-                };
-                (result, None)
-            }
-            // regular expressions string conditionals
-            #[cfg(feature = "regular-expressions")]
-            "regexp" => {
-                use regress::Regex;
-
-                let mut arguments = command.arguments.iter();
-                let source: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                let expression: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-
-                let re = Regex::new(expression).unwrap();
-                let result = re.find(source);
-
-                if let Some(r#match) = result {
-                    if let Some("extract") = arguments.next().map(|argument| argument.0) {
-                        let name: &str = &evaluate_argument(arguments.next().unwrap(), ctx);
-                        let range = r#match
-                            .named_groups()
-                            .find_map(|(key, range)| (key == name).then_some(range));
-                        if let Some(range) = range.flatten() {
-                            (source[range].to_owned(), Some(0))
-                        } else {
-                            eprintln!("No group {name}");
-                            (String::default(), Some(1))
-                        }
-                    } else {
-                        (source[r#match.range].to_owned(), Some(0))
-                    }
-                } else {
-                    eprintln!("No match");
-                    (String::default(), Some(1))
-                }
-            }
-            // Utilities
-            #[cfg(feature = "date")]
-            "date" => {
-                use jiff::{Timestamp, fmt::strtime};
-
-                let mut arguments = command.arguments.iter();
-                let format = if let Some(argument) = arguments.next() {
-                    evaluate_argument(argument, ctx)
-                } else {
-                    Cow::Borrowed("%a %-d %b %Y %T %z")
-                };
-                let now = Timestamp::now();
-                if let Ok(rendered) = strtime::format(&*format, now) {
-                    (rendered, None)
-                } else {
-                    eprintln!("Invalid format {format}");
-                    (String::default(), Some(1))
-                }
-            }
-            // Control flow
-            "if_equal" => {
-                let mut arguments = command.arguments.iter();
-                let first_argument = arguments.next().unwrap();
-                let second_argument = arguments.next().unwrap();
-                let then = arguments.next().unwrap();
-                let equal = evaluate_argument(first_argument, ctx)
-                    == evaluate_argument(second_argument, ctx);
-                let out = if equal {
-                    evaluate_argument(then, ctx).into_owned()
-                } else {
-                    arguments
-                        .next()
-                        .map(|r#else| evaluate_argument(r#else, ctx).into_owned())
-                        .unwrap_or_default()
-                };
-                (out, None)
-            }
-            "if_contains" => {
-                let mut arguments = command.arguments.iter();
-                let item = arguments.next().unwrap();
-                let substring = arguments.next().unwrap();
-                let then = arguments.next().unwrap();
-                let contains =
-                    evaluate_argument(item, ctx).contains(&*evaluate_argument(substring, ctx));
-                let out = if contains {
-                    evaluate_argument(then, ctx).into_owned()
-                } else {
-                    arguments
-                        .next()
-                        .map(|r#else| evaluate_argument(r#else, ctx).into_owned())
-                        .unwrap_or_default()
-                };
-                (out, None)
-            }
-            // TODO WIP. "known programs"
-            command_name @ ("cargo" | "git" | "gh" | "hyperfine" | "jq" | "yq" | "node"
-            | "deno" | "bun" | "sqlite3" | "python" | "npm" | "bat") => {
-                let args = command
-                    .arguments
-                    .iter()
-                    .map(|arg| evaluate_argument(arg, ctx).into_owned())
-                    .filter(|arg| !arg.is_empty())
-                    .collect::<Vec<String>>();
-
-                let (output, result) =
-                    crate::utilities::run_command(command_name.to_owned(), args, None, true, false);
-
-                (output, result.code())
-            }
-            // For constants
-            "literal" | "constant" => {
-                // skip any others
-                let first_argument = command.arguments.first().unwrap();
-                (evaluate_argument(first_argument, ctx).into_owned(), None)
-            }
-            // For conditionally invoking commands
-            "noop" => (String::default(), None),
-            name => {
-                eprintln!("unknown command '{name}'");
-                (String::default(), Some(1))
-            }
-        }
-    }
-}
-
-mod utilities {
-    use std::{env, fs, path::Path};
-
-    pub fn get_environment_variable(name: &str) -> Option<String> {
-        env::vars().find_map(|(n, v)| (n == name).then_some(v))
-    }
-
-    /// `remove_after = true => move`, `remove_after = false => copy`
-    /// TODO if `remove_after`, in some cases can [rename](https://doc.rust-lang.org/std/fs/fn.rename.html) sometimes here
-    pub fn move_copy_file(
-        from: &Path,
-        to: &Path,
-        remove_after: bool,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if from.is_dir() {
-            todo!("copy/move directory");
-        } else if from.is_file() {
-            if let Some(parent) = to.parent() {
-                fs::create_dir_all(parent)?;
-            }
-            let content = fs::read(from)?;
-            fs::write(to, content)?;
-
-            let metadata = fs::metadata(from)?;
-            let permissions = metadata.permissions();
-            fs::set_permissions(to, permissions)?;
-
-            if remove_after {
-                fs::remove_file(from)?;
-            }
-
-            Ok(())
-        } else {
-            Err("Unknown path item to move".into())
-        }
-    }
-
-    /// Reverse <https://howtospell.co.uk/y-to-ies-or-s-plural-rule>
-    pub fn depluralise(on: &str) -> Option<&str> {
-        on.strip_suffix("ies").or_else(|| on.strip_suffix("s"))
-    }
-
-    pub fn run_command(
-        command: String,
-        args: Vec<String>,
-        env: Option<Vec<(String, String)>>,
-        capture_stdout: bool,
-        capture_stderr: bool,
-    ) -> (String, std::process::ExitStatus) {
-        use std::io::{Read, pipe};
-        use std::process::{Command, Stdio};
-
-        let env = env.unwrap_or_default();
-
-        let (stdout, stderr, reader): (Stdio, Stdio, Option<_>) =
-            match (capture_stdout, capture_stderr) {
-                (true, true) => {
-                    // Using pipe we collect both stdout and stderr in order
-                    let (reader, writer) = pipe().expect("could not create pipe");
-                    (
-                        writer
-                            .try_clone()
-                            .expect("could not clone writer pipe")
-                            .into(),
-                        writer.into(),
-                        Some(reader),
-                    )
-                }
-                (true, false) => {
-                    // Shouldn't need pipe here
-                    let (reader, writer) = pipe().expect("could not create pipe");
-                    (writer.into(), Stdio::inherit(), Some(reader))
-                }
-                (false, true) => {
-                    // Shouldn't need pipe here
-                    let (reader, writer) = pipe().expect("could not create pipe");
-                    (Stdio::null(), writer.into(), Some(reader))
-                }
-                (false, false) => (Stdio::inherit(), Stdio::inherit(), None),
-            };
-
-        let mut child = Command::new(command)
-            .args(args)
-            .stdout(stdout)
-            .stderr(stderr)
-            .envs(env)
-            .spawn()
-            .expect("Failed to spawn command");
-
-        if let Some(mut reader) = reader {
-            let mut output = String::new();
-            reader.read_to_string(&mut output).expect("invalid UTF8");
-            let result = child.wait().expect("command not finished");
-            // Remove whitespace from end
-            output.truncate(output.trim_end().len());
-            (output, result)
-        } else {
-            let result = child.wait().expect("command not finished");
-            (String::default(), result)
-        }
-    }
-
-    pub fn separate_numbers(whole_part: &str) -> String {
-        let n: Vec<char> = whole_part.chars().collect();
-        let mut s = String::new();
-        for part in n.rchunks(3).rev() {
-            if !s.is_empty() {
-                s.push(' ');
-            }
-            let chunk: String = part.iter().copied().collect();
-            s.push_str(&chunk);
-        }
-        s
-    }
-
-    pub fn separate_numbers_fract(fract_part: &str) -> String {
-        let n: Vec<char> = fract_part.chars().collect();
-        let mut s = String::new();
-        for part in n.chunks(3) {
-            if !s.is_empty() {
-                s.push(' ');
-            }
-            let chunk: String = part.iter().copied().collect();
-            s.push_str(&chunk);
-        }
-        s
-    }
-}
-
 mod interactive {
     use super::evaluate::{Context, evaluate_statement};
-    use super::parsing::parse_statement;
+    use super::parsing::parsing::parse_statement;
 
     #[derive(Default)]
     pub struct InteractiveState {
